# pkg/api/http/auth/v1/models/logged_in_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// LoggedInDTO logged in d t o
//
// swagger:model LoggedInDTO
type LoggedInDTO struct {

	// refresh token
	RefreshToken string `json:"refresh_token,omitempty"`

	// token
	Token string `json:"token,omitempty"`
}

// Validate validates this logged in d t o
func (m *LoggedInDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this logged in d t o based on context it is used
func (m *LoggedInDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *LoggedInDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *LoggedInDTO) UnmarshalBinary(b []byte) error {
	var res LoggedInDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}