# pkg/api/http/auth/v1/models/profile_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// ProfileDTO profile d t o
//
// swagger:model ProfileDTO
type ProfileDTO struct {

	// active
	Active bool `json:"active,omitempty"`

	// created at
	CreatedAt string `json:"created_at,omitempty"`

	// id
	ID string `json:"id,omitempty"`

	// name
	Name string `json:"name,omitempty"`

	// public key
	PublicKey string `json:"public_key,omitempty"`

	// roles
	Roles []string `json:"roles"`

	// updated at
	UpdatedAt string `json:"updated_at,omitempty"`
}

// Validate validates this profile d t o
func (m *ProfileDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this profile d t o based on context it is used
func (m *ProfileDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ProfileDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ProfileDTO) UnmarshalBinary(b []byte) error {
	var res ProfileDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}