# pkg/api/http/auth/v1/models/register_d_t_o.go
// Code generated by go-swagger; DO NOT EDIT.

package models

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// RegisterDTO register d t o
//
// swagger:model RegisterDTO
type RegisterDTO struct {

	// password
	Password string `json:"password,omitempty"`

	// username
	Username string `json:"username,omitempty"`
}

// Validate validates this register d t o
func (m *RegisterDTO) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this register d t o based on context it is used
func (m *RegisterDTO) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *RegisterDTO) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *RegisterDTO) UnmarshalBinary(b []byte) error {
	var res RegisterDTO
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}